# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "N100 Financial Intelligence Platform",
    "description": "FastAPI Backend for N100 Financial Intelligence Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sectors": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "summary": "Get Sector Summaries",
        "description": "Retrieve sector aggregated statistics including company count, median ROE, median P/E, and median D/E.",
        "operationId": "get_sectors_api_v1_sectors_get",
        "responses": {
          "200": {
            "description": "Sector summary list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SectorSummaryItem"
                  },
                  "type": "array",
                  "title": "Response Get Sectors Api V1 Sectors Get"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/sectors/{sector}/companies": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "summary": "Get Companies by Sector",
        "description": "Retrieve list of companies and latest financial KPIs belonging to a specific sector.",
        "operationId": "get_sector_companies_api_v1_sectors__sector__companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of companies in sector returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectorCompanyItem"
                  },
                  "title": "Response Get Sector Companies Api V1 Sectors  Sector  Companies Get"
                }
              }
            }
          },
          "404": {
            "description": "Sector not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/market-cap/{ticker}": {
      "get": {
        "tags": [
          "Valuation"
        ],
        "summary": "Get Historical Valuation Multiples",
        "description": "Retrieve historical valuation multiples (P/E, P/B, EV/EBITDA, Dividend Yield) from 2019 to 2024 for a company.",
        "operationId": "get_company_market_cap_valuation_api_v1_market_cap__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical valuation data returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyValuationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company ticker not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyValuationResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "description": "Company ticker symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Official company name"
          },
          "historical_valuation": {
            "items": {
              "$ref": "#/components/schemas/ValuationYearRecord"
            },
            "type": "array",
            "title": "Historical Valuation",
            "description": "Chronological annual valuation records (2019 to 2024)"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "company_name",
          "historical_valuation"
        ],
        "title": "CompanyValuationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SectorCompanyItem": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id",
            "description": "Unique company ticker symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Official company name"
          },
          "sector": {
            "type": "string",
            "title": "Sector",
            "description": "Sector classification"
          },
          "roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe",
            "description": "Latest Return on Equity %"
          },
          "roce": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roce",
            "description": "Latest Return on Capital Employed %"
          },
          "debt_to_equity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt To Equity",
            "description": "Latest Debt to Equity ratio"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio",
            "description": "Latest Price to Earnings ratio"
          },
          "net_profit_margin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Profit Margin",
            "description": "Latest Net Profit Margin %"
          },
          "latest_kpis": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Kpis",
            "description": "Consolidated latest financial KPIs"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "company_name",
          "sector"
        ],
        "title": "SectorCompanyItem"
      },
      "SectorSummaryItem": {
        "properties": {
          "sector": {
            "type": "string",
            "title": "Sector",
            "description": "Name of the industry sector"
          },
          "company_count": {
            "type": "integer",
            "title": "Company Count",
            "description": "Number of companies in this sector"
          },
          "median_roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Roe",
            "description": "Median Return on Equity % across sector"
          },
          "median_pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Pe",
            "description": "Median Price to Earnings ratio across sector"
          },
          "median_de": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median De",
            "description": "Median Debt to Equity ratio across sector"
          }
        },
        "type": "object",
        "required": [
          "sector",
          "company_count"
        ],
        "title": "SectorSummaryItem"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValuationYearRecord": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period",
            "description": "Financial period/year (e.g. 2019, 2024)"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio",
            "description": "Price to Earnings ratio (P/E)"
          },
          "pb_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb Ratio",
            "description": "Price to Book ratio (P/B)"
          },
          "ev_ebitda": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ev Ebitda",
            "description": "EV to EBITDA multiple"
          },
          "dividend_yield": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dividend Yield",
            "description": "Dividend Yield %"
          },
          "market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap",
            "description": "Market Capitalization in Crores"
          },
          "enterprise_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enterprise Value",
            "description": "Enterprise Value in Crores"
          }
        },
        "type": "object",
        "required": [
          "period"
        ],
        "title": "ValuationYearRecord"
      }
    }
  }
}